Convert arithmetic scalars to new signature registry
By using signature matching of arithmetic scalars, arithmetic data type
selection is based on precedence and thus fixes existing issues for
floating point arithmetics.
Operations on non-floating point column types with a floating point
literal will result in a floating point result now.

Fixes #9652

+++ server/src/test/java/io/crate/analyze/expressions/ExpressionAnalyzerTest.java
@@ -330,7 +330,7 @@ public class ExpressionAnalyzerTest extends CrateDummyClusterServiceUnitTest {
             symbol2.arguments().get(1),
             isFunction(
                 ImplicitCastFunction.NAME,
-                List.of(DataTypes.INTEGER, DataTypes.STRING)
+                List.of(DataTypes.DOUBLE, DataTypes.STRING)
             )
         );
     }
